Root Component
The `Root` component has been enhanced so that it displays different
components (i.e. `App` or `Guest`) based on whether the user is
authenticated.

It now accepts the following two `props`:

- isLoggedIn -> Boolean
As it name implies,  this prop will be used to indicate to the
component whether the current user is authenticated and will be
determined from the app's redux store. If the user is authenticated,
the `Root` component will render the `App` component and display the
pages for authenticated users (ex. Profile, Complex Pages, etc...).

On the other hand if the user is not logged in, before displaying the
`Guest` component the `Root` component will first try to authenticate
the user using cookies (note that this is not yet implemented as
expected since backend is not in scope).

- login -> Function
This callback is meant to be invoked with the users object when the
user is authenticated from the `Root` component. When invoked it will
update the `session` redux namespace (used to store the ID of the logged
in user) & the `user` redux namespace (used to store the info of all
users).