Create thunk for user authentication
This thunk can be used to authenticate a user using either his
credentials or HTTP Cookies. When the user is successfully
authenticated a resolved promise is returned and the following two
actions are dispatched to the app's redux store:

  > USER_ADD: Caching the details of authenticated user.
  > SESSION_LOGIN: Authenticating the user.

When authentication fails a promise rejected with an error object
describing what went wrong is returned. Based on the cause of the
authentication failure the following errors can be returned:

  > AuthError:
    Indicating that either the provided credentials or HTTP Cookies
    are invalid. This error is thrown when the HTTP Status is 401.

  > ServerError:
    Indicating that something went wrong on the server. This error
    is thrown when the HTTP Status is neither 200 or 401.

  > Error:
    Thrown when something went wrong from the client side.