Refactor Login Thunk unit tests
In the Successful Authentication unit test, to verify whether the
right Action Creators are used & whether they are invoked using the
correct arguments, the actions dispatched to redux store are being
asserted.

Using this approach however is wrong since we are testing code that
the Login Thunk is dependent on.

In this commit the unit test is refactored so that the USER_ADD and
SESSION_LOGIN Action Creators are mocked to instead of returning
the action they return an object storing info about the way the
Action Creators are invoked.