Use redux thunk
When the user first opens the application an authentication request
(using http cookies) is made so that if the user's token is still
valid, he won't be required to re-log. Currently this request is being
made in the `Root` component however when the Login Page was being
implemented it was seen that the following logic for logging the user
in was being duplicated:

- Fetch Request.
- Dispatch of `USER_ADD` & `SESSION_ADD` redux action on successful
  authentication.
- Error on unsuccessful authentication.

For this reason it was decided to place the Login logic inside a redux
thunk.

In addition to the current implementation, the Login functionality now
also can return the following three distinct errors when the user fails
to authenticate:
- AuthError: Thrown when the authentication failed (i.e. HTTP Status
  401).
- ServerError: Thrown when the HTTP Status is not 401.
- Error: Thrown when there is a problem from the client side.

During the development of this PR `material-ui` v1 was also released
and therefore the version was bumped.