Allow a user to visit any page using URL
With the current implementation of the BottomNavigation component,
the user was unable to visit a non-root page via the URL since the
logic in the BottomNavigation component would change the URL to
'/'.

In this commit the BottomNavigation was updated to fix this issue.
In addition to this it was enhanced to highlight a
BottomNavigationAction if a child page of the page that the
BottomNavigationAction links to, is being displayed for example a
BottomNavigationAction linking to '/' would be highlighted
if the URL is '/messages/new'.